refactor: ♻️ resolve all Qodana code quality issues
- Add @NonNull annotations to method parameters overriding @NotNull
- Add null checks and Objects.requireNonNull() for potentially null returns
- Replace printStackTrace() with proper ComponentLogger.error() calls
- Remove unused imports across multiple files
- Convert field constants to local variables where appropriate
- Add @SuppressWarnings for unavoidable unchecked warnings
- Fix stream API calls (collect(toList()) → toList())
- Replace size comparisons with isEmpty()
- Fix non-regex replaceAll() to use replace()
- Use parameterized logging to avoid string concatenation

Fixes 125+ code quality issues identified in Qodana report while maintaining all functionality.
The main remaining issues are Message / Component related which will be fixed once we add translation.
Hopefully very soon.

Now we don't set a baseline anymore for better reporting.

+++ src/main/java/net/buildtheearth/buildteamtools/utils/BlockLocation.java
@@ -27,7 +27,7 @@ public class BlockLocation {
      * @return Bukkit Location object for this position
      */
     public Location getLocation(World world){
-        
-        return new Location(world, (double)x, (double)y, (double)z);
+
+        return new Location(world, x, y, z);
     }
 }
